Moved initializecms into the cms group and modified test cases.
