+++ test/distributed/org/apache/cassandra/distributed/upgrade/ClusterMetadataUpgradeHarryTest.java
@@ -117,8 +117,8 @@ public class ClusterMetadataUpgradeHarryTest extends UpgradeTestBase
                 assertEquals(0, res.length);
             });
 
-            cluster.get(1).nodetoolResult("initializecms").asserts().success();
-            cluster.get(1).nodetoolResult("reconfigurecms", "3").asserts().success();
+            cluster.get(1).nodetoolResult("cms", "initialize").asserts().success();
+            cluster.get(1).nodetoolResult("cms", "reconfigure", "3").asserts().success();
             cluster.schemaChange(withKeyspace("create table %s.xyz (id int primary key)"));
             stopLatch.decrement();
             harryRunner.get().get();
