+++ test/distributed/org/apache/cassandra/distributed/upgrade/ClusterMetadataUpgradeTest.java
@@ -48,14 +48,14 @@ public class ClusterMetadataUpgradeTest extends UpgradeTestBase
             cluster.schemaChange("CREATE TABLE " + KEYSPACE + ".tbl (pk int, ck int, v int, PRIMARY KEY (pk, ck))");
         })
         .runAfterClusterUpgrade((cluster) -> {
-            cluster.get(1).nodetoolResult("initializecms").asserts().success();
+            cluster.get(1).nodetoolResult("cms","initialize").asserts().success();
             cluster.forEach(i ->
             {
                 // The cast is unpleasant, but safe to do so as the upgraded instance is running the current version.
                 assertFalse("node " + i.config().num() + " is still in MIGRATING STATE",
                             ClusterUtils.isMigrating((IInvokableInstance) i));
             });
-            cluster.get(2).nodetoolResult("reconfigurecms", "3").asserts().success();
+            cluster.get(2).nodetoolResult("cms", "reconfigure", "3").asserts().success();
             cluster.schemaChange(withKeyspace("create table %s.xyz (id int primary key)"));
             cluster.forEach(i -> {
                 Object [][] res = i.executeInternal("select host_id from system.local");
